Added Logitech Precission controller (incomplete)
The controller doesn't have thumbsticks. Bytes 0 and 1 "simulate" a
stick from the d-pad, so you could use that for main stick. The problem
is that there is nowhere to "plug in" the c-stick. In this revision,
d-pad is disabled, c-stick is on unused (null) byte, i.e. forced to
top-left corner, and normal stick is controlled by real d-pad.
I'm going to figure out some workaround later.