[controller_ini_to_array] Fixed invalid buffer size
The file is opened in text-mode, meaning that ftell doesn't give you
"amount of characters that you can read", but rather a "value that can
be used with fseek SEEK_SET". That value is usually a binary offset
within the file, which is obviously bigger than text ones (all the
newlines count twice due to being \r\n instead of \n).
Due to this, the buffer is allocated with more space than there is data
to read in the file, which results in junk data between end of the file
and the NULL-byte at the end of the buffer.
The junk data only affect us in two cases:
1) junk contains an entry which isn't in the ini file. Example, ini file
doesn't have DPAD entry (so it defaults to 0), but junk area contains
"\nDPAD=1\n".
2) If our input file doesn't have a newline at the very end, the last
value could be corrupted. Example: "DigitalLR=1\n0", if you remove that
newline it will become "DigitalLR=10" which is not good.

Switching to binary-mode fixes this obscure bug. (AFAIK fatfs only has
binary mode, so no changes needed in kernel)